[Pytorch][test] Fixing Pt-1.9 SM failures (#1976)
* Fixing Pt-19 SM failures

* Change file to the older commit(https://github.com/aws/deep-learning-containers/blob/b87590af021a4c3913b3669fa5f9fc10fbebe4e8/test/sagemaker_tests/pytorch/training/resources/smdataparallel/smdataparallel_throughput.py)

* Adding version split based on SMDDP team's suggestion to run smdataparallel_throughput_post_ptbackend

* Removed incorrect line to test normal working

* Revert temp changes
